+++ test/models/v2/channel.spec.ts
@@ -8,7 +8,7 @@ import { Message } from '../../../src/models/v2/message';
 import { Servers } from '../../../src/models/v2/servers';
 import { Server } from '../../../src/models/v2/server';
 
-import { serializeInput, assertBindings, assertDescription, assertExtensions } from './assert-mixins';
+import { serializeInput, assertBindings, assertDescription, assertExtensions } from './utils';
 
 import type { v2 } from '../../../src/interfaces';
