+++ test/models/v2/external-docs.spec.ts
@@ -1,6 +1,6 @@
 import { ExternalDocumentation } from '../../../src/models/v2/external-docs';
 
-import { serializeInput, assertDescription, assertExtensions } from './assert-mixins';
+import { serializeInput, assertDescription, assertExtensions } from './utils';
 
 import type { v2 } from '../../../src/interfaces';
