+++ test/models/v2/license.spec.ts
@@ -1,6 +1,6 @@
 import { License } from '../../../src/models/v2/license';
 
-import { serializeInput, assertExtensions } from './assert-mixins';
+import { serializeInput, assertExtensions } from './utils';
 
 import type { v2 } from '../../../src/interfaces';
